hv: move sbuf_share_setup() and sbuf_reset() to common
acrntrace and hvlog depends on sbuf, as they are going to be included
in release versions, sbuf also need to be included.

Tracked-On: #8805
